+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All changes to this project will be noted here.
 
+## [1.0.0] - 2024-01-20
+
+### Added
+
+Optional features for this plug-in have been added
+
+- `type` - "nextjs" - Application-specific directories Only supported by Next.js
+- `directory` - Directory to replace obfuscated class names
+
+### Revision
+
+- Fixed a problem that prevented the plugin from working properly
+
 ## [0.1.0] - 2024-01-20
 
 ### Added
@@ -9,3 +22,4 @@ All changes to this project will be noted here.
 - First release
 
 [0.1.0]: https://www.npmjs.com/package/postcss-classname-obfuscator/v/0.1.0
+[0.1.0]: https://www.npmjs.com/package/postcss-classname-obfuscator/v/1.0.0